ipfixprobe - introduce process ICMP plugin
