+++ src/plugins/process/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(basicplus)
 add_subdirectory(bstats)
+add_subdirectory(icmp)
